[bug-fix] replace semantic accordion with amp accordion // #161
