+++ reframe/browser/browserInit.js
@@ -1,7 +1,6 @@
 import browserConfig from '@brillout/browser-config';
 
 import './jquery-global'; // see https://stackoverflow.com/a/39820703/1855917
-import './thirdparty/semantic-ui-2.3.2/accordion.min';
 
 initBrowser();
