+++ reframe/views/utils/Accordion/AccordionAmpFix.css
@@ -1,0 +1,6 @@
+div.i-amphtml-accordion-header {
+  border: 0;
+  padding-right: 0;
+  background-color: transparent;
+  outline: 0;
+}
