+++ reframe/views/utils/Accordion/AccordionBadgerIcon.css
@@ -1,4 +1,5 @@
 /* CSS copied and adapted from https://github.com/stuartjnelson/badger-accordion */
+
 .badger-accordion__header-icon {
   float: right;
   position: absolute;
